Cleaning up VCK5000 examples
