+++ reference_designs/vck5000/vector_scalar_kernel/Makefile
@@ -10,9 +10,9 @@ include ../makefile-common
 
 ACDC_AIE = $(dir $(shell which aie-opt))/..
 
-targetname = vectorScalar
+SHELL := /bin/bash
 
-all: build/final.xclbin build/insts.txt
+all: build/aie.mlir vck5000 run
 
 scale.o: scale.cc
 	xchesscc -p me -P ${VITIS_ROOT}/aietools/data/versal_prod/lib/ -L ${VITIS_ROOT}/aietools/data/versal_prod/lib/ -c ./scale.cc
@@ -21,11 +21,6 @@ build/aie.mlir: aie2.py
 	mkdir -p ${@D}
 	python3 $< > $@
 
-build/final.xclbin: build/aie.mlir build/scale.o
-	mkdir -p ${@D}
-	cd ${@D} && aiecc.py --aie-generate-cdo --no-compile-host --xclbin-name=${@F} \
-				--aie-generate-ipu --ipu-insts-name=insts.txt $(<:%=../%)
-
 vck5000: build/aie.mlir scale.o
 	aiecc.py --link_against_hsa --host-target=x86_64-amd-linux-gnu build/aie.mlir \
 						-I/opt/xaiengine/include \
@@ -33,26 +28,15 @@ vck5000: build/aie.mlir scale.o
             -I${ROCM_ROOT}/include \
 						-L/opt/xaiengine/lib \
             -L/lib/x86_64-linux-gnu/ \
-						test_vck5000.cpp \
+						test.cpp \
 						$(ACDC_AIE)/runtime_lib/x86_64-hsa/test_lib/src/test_library.cpp \
             ${ROCM_ROOT}/lib/libhsa-runtime64.so.1.9.0 \
 						-Wl,-R/opt/xaiengine/lib \
             -Wl,-rpath,${ROCM_ROOT}/lib \
 						-Wl,--whole-archive -Wl,--no-whole-archive -lstdc++ -ldl -lelf -o test.elf
 
-${targetname}.exe: test.cpp
-	rm -rf _build
-	mkdir -p _build
-	cd _build && ${powershell} cmake .. -DTARGET_NAME=${targetname}
-	cd _build && ${powershell} cmake --build . --config Release
-ifeq "${powershell}" "powershell.exe"
-	cp _build/${targetname}.exe $@
-else
-	cp _build/${targetname} $@ 
-endif
-
-run: ${targetname}.exe build/final.xclbin build/insts.txt 
-	${powershell} ./$< -x build/final.xclbin -i build/insts.txt -k MLIR_AIE
+run: test.elf
+	./test.elf
 
 clean:
-	rm -rf build _build inst aie.mlir.prj core_* test.elf scale.o*
+	rm -rf build _build inst aie.mlir.prj core_* test.elf *.o
